backend/requirements.txt
"""Requirements for the backend application."""
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AI/ML
google-generativeai==0.4.1

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
python-json-logger==2.0.7

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.1
